knowledge-service/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
pyyaml==6.0.2
python-frontmatter==1.1.0
watchdog==4.0.0
httpx==0.27.0
pydantic==2.9.0
pytest>=8.0.0
pytest-asyncio>=0.24.0
